+++ CHANGES.rst
@@ -28,6 +28,7 @@ Minor Changes
 - test: prepare fixtures to refactor 'test_PackageSigner.py'
 - test: refactor 'test_PackageSigner.py' & cover all branches
 - test: cover input combinations for ``(a)test_timestamp``
+- test: import ``hypothesis`` in 'conftest.py'
 - refactor: remove unnecessary variable alias